too much errors

+++ ft_itoa.c
@@ -6,17 +6,17 @@
 /*   By: cde-la-r <cde-la-r@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/15 09:10:03 by cde-la-r          #+#    #+#             */
-/*   Updated: 2023/09/28 00:02:27 by cde-la-r         ###   ########.fr       */
+/*   Updated: 2023/09/28 19:21:48 by cde-la-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-char	*ft_itoa(int n)
+char	*ft_itoa(long long n)
 {
-	int				l;
-	unsigned int	num;
-	char			*r;
+	int					l;
+	unsigned long long	num;
+	char				*r;
 
 	l = (n == 0);
 	num = n * (n >= 0) - n * (n < 0);
