+++ ft_itoa_base.c
@@ -6,18 +6,18 @@
 /*   By: cde-la-r <cde-la-r@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/27 17:47:56 by cde-la-r          #+#    #+#             */
-/*   Updated: 2023/09/28 00:01:03 by cde-la-r         ###   ########.fr       */
+/*   Updated: 2023/09/28 19:04:07 by cde-la-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-char	*ft_itoa_base(unsigned int n, const char *base)
+char	*ft_itoa_base(unsigned long long n, const char *base)
 {
-	int		nbase;
-	int		l;
-	int		tmp;
-	char	*r;
+	int						nbase;
+	int						l;
+	unsigned long long		tmp;
+	char					*r;
 
 	nbase = ft_strlen(base);
 	l = (n == 0);
